requirements.txt
sqlalchemy
fastapi
uvicorn
databases
pydantic
passlib[bcrypt]
python-jose